% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_normal}
\alias{standard_normal}
\title{Standard normally distributed values}
\format{
A data frame with 100 rows and 1 column:
\describe{
  \item{Value}{Normally distributed values.}
}
}
\usage{
standard_normal
}
\description{
Table with normally distributed values
with mean = 0 and standard deviation = 1.
}
\keyword{datasets}
